refactor: #121 UIApplication 확장 분리
