+++ ByeBoo-iOS/ByeBoo-iOS/Presentation/Extension/UIApplication+.swift
@@ -1,0 +1,27 @@
+//
+//  UIApplication+.swift
+//  ByeBoo-iOS
+//
+//  Created by APPLE on 7/16/25.
+//
+
+import UIKit
+
+extension UIApplication {
+    func topViewController(base: UIViewController? = UIApplication.shared
+        .connectedScenes
+        .compactMap { $0 as? UIWindowScene }
+        .flatMap { $0.windows }
+        .first(where: { $0.isKeyWindow })?
+        .rootViewController
+    ) -> UIViewController? {
+        if let nav = base as? UINavigationController {
+            return topViewController(base: nav.visibleViewController)
+        } else if let tab = base as? UITabBarController {
+            return topViewController(base: tab.selectedViewController)
+        } else if let presented = base?.presentedViewController {
+            return topViewController(base: presented)
+        }
+        return base
+    }
+}
