+++ ByeBoo-iOS/ByeBoo-iOS/Presentation/Feature/Quest/View/QuestCheckViewController.swift
@@ -272,24 +272,4 @@ extension QuestCheckViewController: UICollectionViewDataSource {
         
         return headerView
     }
-    
-}
-
-extension UIApplication {
-    func topViewController(base: UIViewController? = UIApplication.shared
-        .connectedScenes
-        .compactMap { $0 as? UIWindowScene }
-        .flatMap { $0.windows }
-        .first(where: { $0.isKeyWindow })?
-        .rootViewController
-    ) -> UIViewController? {
-        if let nav = base as? UINavigationController {
-            return topViewController(base: nav.visibleViewController)
-        } else if let tab = base as? UITabBarController {
-            return topViewController(base: tab.selectedViewController)
-        } else if let presented = base?.presentedViewController {
-            return topViewController(base: presented)
-        }
-        return base
-    }
 }
